Добавляю файл с данными
