+++ js/data.js
@@ -1,0 +1,22 @@
+const messageList = [
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.'
+];
+
+const description = [
+  'Мне понравилась эта фотография, потому что она передаёт чувства и эмоции присутствующих на ней людей.',
+  'прекрасную атмосферу летнего вечера.',
+  'Всё отлично!'
+];
+
+const namesList =  [
+  'Иван',
+  'Роман',
+  'Мария',
+  'Светлана',
+  'Федор',
+  'Екатерина',
+  'Максим'
+];
+
+export {messageList, description, namesList};
