+++ js/main.js
@@ -1,3 +1,4 @@
 import './util.js';
 import './random-comment.js';
 import './decription-photo.js';
+import './data.js';
